requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# AWS dependencies (for Kinesis)
# NOTE: boto3 is included in the AWS Lambda runtime and is excluded 
# from the deployment package by the build_lambda_package.py script.
boto3>=1.28.0
